pyproject.toml
[tool.poetry]
name = "send-mails"
version = "0.1.0"
description = "sending mails automatically via gmail"
authors = ["mxmlnwbr <maximilian.weber@bluewin.ch>"]
readme = "README.md"
packages = [{include = "send_mails"}]

[tool.poetry.dependencies]
python = "^3.10"
python-decouple = "^3.8"
python-dotenv = "^1.0.0"
openpyxl = "^3.1.2"
pandas = "^2.1.3"
tqdm = "^4.66.1"
secure-smtplib = "^0.1.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
